Renamed remp variables, used test fixture, moved to serializers.py
